backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
openpyxl
asyncpg==0.29.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
